% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_implied_volatility.R
\name{append_implied_volatility}
\alias{append_implied_volatility}
\title{Append European Option Implied Volatilities to DataFrame}
\usage{
append_implied_volatility(df, risk_free_rate = 0.01,
  initial_volatility_guess = 0.5)
}
\arguments{
\item{df}{Dataframe with columns named type, value, underlying, strike, and maturity}

\item{risk_free_rate}{Risk-free rate (number, default: 0.01)}

\item{initial_volatility_guess}{Initial guess for the volatility (number, default: 0.5)}
}
\value{
implied volatility of a European Option appended (dataframe)
}
\description{
\code{append_implied_volatility} returns a dataframe with European Option implied volatilities
}
\details{
This function computes the volatilty of a European Option for each row
in the dataframe that is passed to it. It will do so using the
\code{compute_implied_volatility} function in this package, which is a
wrapper around the \code{EuropeanOptionImpliedVolatility} function from the
\code{RQUantLib} package.


The columns in the dataframe should be as follows:
- \code{type}: One of the two values: 'call' or 'put' (string)
- \code{value}: Value of the option (number)
- \code{underlying}: Current price of the underlying (number)
- \code{strike}: Strike price of the option (number)
- \code{maturity}: Time to maturity in fractional years (number)

TODO: Model/financial restrictions on the parameters should be made clear here...

The \code{EuropeanOptionImpliedVolatility} function asks for a
\code{dividend_yield} which we assume to be 0 as this is not a stock.

Fore more information about RQuantLib:
https://cran.r-project.org/web/packages/RQuantLib/RQuantLib.pdf

For more information about QuantLib:
http://quantlib.org/index.shtml
}
\author{
John Dole <jdoleiv@gmail.com>
}
\seealso{
\code{EuropeanOptionImpliedVolatility}

\code{append_implied_volatility}
}

