% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_and_graph_implied_volatility.R
\name{compute_and_graph_implied_volatility}
\alias{compute_and_graph_implied_volatility}
\title{Append and Graph Implied Volatility}
\usage{
compute_and_graph_implied_volatility(df, risk_free_rate = 0.01,
  initial_volatility_guess = 0.5)
}
\arguments{
\item{df}{Dataframe with option data (see details)}

\item{risk_free_rate}{Risk-free rate (number, default: 0.01)}

\item{initial_volatility_guess}{Initial guess volatility guess (default: 0.5)}
}
\description{
\code{compute_and_append_implied_volatility} returns a dataframe with the implied volatility appended and graphs it
}
\details{
This function uses the functions \code{append_implied_volatility} and
\code{graph_implied_volatility} from this package to compute and graph
the implied volatility for European Options, respectively.
}
\author{
John Dole <jdoleiv@gmail.com>
}
\seealso{
\code{append_implied_volatility}

\code{compute_implied_volatility}

\code{graph_implied_volatility}
}

