% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_implied_volatility.R
\name{compute_implied_volatility}
\alias{compute_implied_volatility}
\title{Compute European Option Implied Volatility}
\usage{
compute_implied_volatility(type, value, underlying, strike, maturity,
  risk_free_rate = 0.01, initial_volatility_guess = 0.5)
}
\arguments{
\item{type}{One of the two values: 'call' or 'put' (string)}

\item{value}{Value of the option (number)}

\item{underlying}{Current price of the underlying (number)}

\item{strike}{Strike price of the option (number)}

\item{maturity}{Time to maturity in fractional years (number)}

\item{risk_free_rate}{Risk-free rate (number, default: 0.01)}

\item{initial_volatility_guess}{Initial guess for the volatility (number, default: 0.5)}
}
\value{
implied volatility of a European Option (number)
}
\description{
\code{compute_implied_volatility} returns the implied volatility of a European Option
}
\details{
This function is a wrapper around the \code{EuropeanOptionImpliedVolatility}
function from the \code{RQUantLib} package. This function is used internally,
in this package, to compute a dataframe with implied volatitilies for each row.


TODO: Model/financial restrictions on the parameters should be made clear here...

The \code{EuropeanOptionImpliedVolatility} function asks for a
\code{dividend_yield} which we assume to be 0 as this is not a stock.

Fore more information about RQuantLib:
https://cran.r-project.org/web/packages/RQuantLib/RQuantLib.pdf

For more information about QuantLib:
http://quantlib.org/index.shtml
}
\author{
John Dole <jdoleiv@gmail.com>
}
\seealso{
\code{EuropeanOptionImpliedVolatility}
}

