% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_implied_volatility.R
\name{graph_implied_volatility}
\alias{graph_implied_volatility}
\title{Graph Implied Volatility}
\usage{
graph_implied_volatility(df)
}
\arguments{
\item{df}{Dataframe with \code{Strike} and \code{ImpliedVolatility} columns}
}
\description{
\code{graph_implied_volatility} graphs the implied volatility in a dataframe
}
\details{
This function receives as input a dataframe that contains at least the columns
\code{Strike} and \code{ImpliedVolatility}. It primarily intended to be used
with the output dataframe from the \code{append_implied_volatility} function.
}
\author{
John Dole <jdoleiv@gmail.com>
}
\seealso{
\code{append_implied_volatility}

\code{compute_implied_volatility}
}

